added styling
