+++ friends/src/components/Form/FormStyle.css
@@ -1,0 +1,44 @@
+.form {
+    margin: 40px 20px 0px 20px;;
+    display: flex;
+    width: 100%;
+    max-width: 500px;
+    justify-content: center;
+}
+
+.inputs {
+    display: flex;
+    flex-flow: column;
+}
+
+input {
+    margin: 10px;
+    border: none;
+    border: solid black 1px;
+    height: 20px;
+    border-radius: 2px;
+    padding-left: 5px;
+}
+
+.submitbutton {
+    border: none;
+    border: solid black 2px;
+    background: black;
+    color: white;
+    font-size: 20px;
+    width: 140px;
+    margin: 0px 10px;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.notdonebutton {
+    border: none;
+    border: solid black 2px;
+    background: white;
+    color: black;
+    font-size: 20px;
+    width: 140px;
+    margin: 0px 10px;
+    border-radius: 4px;
+}
