+++ friends/src/components/FriendsList/FriendsListStyle.css
@@ -1,0 +1,34 @@
+.container {
+    display: flex;
+    flex-flow: column;
+    align-items: center;
+}
+
+.friendCard {
+    display: flex;
+    flex-flow: column;
+    align-items: flex-start; 
+    border: black 2px solid;
+    width: 100%;
+    max-width: 300px;
+    margin: 10px 0;
+    padding: 5px;
+    border-radius: 4px;
+    background-color: black;
+}
+
+.name {
+    font-size: 20px;
+    font-weight: bold;
+    color: white;
+}
+
+.age {
+    margin-top: 4px;
+    color: white;
+}
+
+.email {
+    margin-top: 4px;
+    color: white;
+}
